+++ app/models/randomization_scheme.rb
@@ -5,6 +5,7 @@ class RandomizationScheme < ActiveRecord::Base
 
   # Constants
   MAX_LISTS = 128
+  ALGORITHMS = [['Permuted-Block Algorithm', 'permuted-block'], ['Minimisation Algorithm', 'minimisation']]
 
   # Concerns
   include Searchable, Deletable
@@ -30,25 +31,7 @@ class RandomizationScheme < ActiveRecord::Base
   # Model Methods
 
   def add_missing_lists!(current_user)
-    list_option_ids = []
-
-    if self.number_of_lists > 0 and self.number_of_lists < MAX_LISTS
-      if self.stratification_factors.count == 1
-        list_option_ids = self.stratification_factors.first.option_hashes.collect{|i| [i]}
-      else
-        list_option_ids = self.stratification_factors.collect{|sf| sf.option_hashes}.inject(:product)
-      end
-    end
-
-    list_option_ids.each do |option_hashes|
-      unless self.find_list_by_option_hashes(option_hashes)
-        stratification_factor_option_ids = option_hashes.collect{ |oh| oh[:stratification_factor_option_id] }
-        options = self.stratification_factor_options.where(id: stratification_factor_option_ids)
-        extra_options = option_hashes.select{ |oh| oh[:extra] }
-
-        self.lists.create(project_id: self.project_id, user_id: current_user.id, options: options, extra_options: extra_options)
-      end
-    end
+    RandomizationAlgorithm.for(self).add_missing_lists!(current_user)
   end
 
   def generate_lists!(current_user)
@@ -59,52 +42,16 @@ class RandomizationScheme < ActiveRecord::Base
     true
   end
 
-  def find_list_by_option_hashes(option_hashes)
-    criteria_pairs = option_hashes.collect{|h| [h[:stratification_factor_id], (h[:stratification_factor_option_id] || h[:site_id])]}
-    self.find_list_by_criteria_pairs(criteria_pairs)
-  end
-
   def find_list_by_criteria_pairs(criteria_pairs)
-    list = nil
-    self.lists.each do |l|
-      if l.criteria_match?(criteria_pairs)
-        list = l
-        break
-      end
-    end
-    list
+    RandomizationAlgorithm.for(self).find_list_by_criteria_pairs(criteria_pairs)
   end
 
-  def randomize_subject_to_list!(subject, list, current_user)
-    # Find next randomization in list
-    randomization = list.randomizations.where(subject_id: nil).order(:position).first
-
-    # Expand lists by another block group
-    unless randomization
-      max_needed_block_group = [self.next_block_group - 1, list.next_block_group].max
-      generate_next_block_group_up_to!(current_user, max_needed_block_group)
-      randomization = list.randomizations.where(subject_id: nil).order(:position).first
-    end
-
-    # Add subject to randomization list
-    randomization.add_subject!(subject, current_user) if randomization
-    randomization
-  end
-
-  def generate_next_block_group_up_to!(current_user, block_group)
-    multipliers = self.block_size_multipliers.collect{|m| [m.value] * m.allocation }.flatten
-    arms        = self.treatment_arms.collect{|arm| [arm.id] * arm.allocation }.flatten
-    self.lists.each do |list|
-      list.generate_all_block_groups_up_to!(current_user, block_group, multipliers, arms)
-    end
-  end
-
-  def next_block_group
-    (randomizations.pluck(:block_group).max || 0) + 1
+  def randomize_subject_to_list!(subject, list, current_user, criteria_pairs)
+    RandomizationAlgorithm.for(self).randomize_subject_to_list!(subject, list, current_user, criteria_pairs)
   end
 
   def number_of_lists
-    self.stratification_factors.collect{ |sf| sf.option_hashes.count }.inject(:*).to_i
+    RandomizationAlgorithm.for(self).number_of_lists
   end
 
   def minimum_block_size
@@ -117,6 +64,26 @@ class RandomizationScheme < ActiveRecord::Base
 
   def has_randomized_subjects?
     self.randomizations.where.not(subject_id: nil).count > 0
+  end
+
+  def algorithm_name
+    if algorithm = ALGORITHMS.select{|name, value| value == self.algorithm}.first
+      algorithm[0]
+    else
+      "No Algorithm Selected"
+    end
+  end
+
+  def permuted_block?
+    self.algorithm == 'permuted-block'
+  end
+
+  def minimisation?
+    self.algorithm == 'minimisation'
+  end
+
+  def randomization_error_message
+    RandomizationAlgorithm.for(self).randomization_error_message
   end
 
   private