+++ config/initializers/app_config.rb
@@ -6,3 +6,5 @@ require 'uri'
 # In lib directory
 require 'date_and_time_parser'
 require 'valuables'
+# In models directory
+require 'randomization_algorithm/randomization_algorithm'