+++ db/migrate/20150713175117_add_algorithm_to_randomization_scheme.rb
@@ -1,0 +1,5 @@
+class AddAlgorithmToRandomizationScheme < ActiveRecord::Migration
+  def change
+    add_column :randomization_schemes, :algorithm, :string, null: false, default: 'permuted-block'
+  end
+end