+++ test/models/randomization_characteristic_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class RandomizationCharacteristicTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end